refactor(gamePlaying): 게임 진행 중 각 말의 차례를 나타내는 상태 객체 추상화
- 상태 객체들을 final class 로 변경
